fix monday update bug - final file in index.js - ready for PR
